+++ src/views/resources/UpdatesListView.vue
@@ -2,10 +2,10 @@
   <div>
     <PageSectionTitle :pageTitle="title" />
     <v-row>
-      <v-col :cols="`${mdAndUp ? 2 : 12}`">
+      <v-col cols="12" md="2">
         <UpdatesTable></UpdatesTable>
       </v-col>
-      <v-col cols="10">
+      <v-col cols="12" md="10">
         <!-- Suspense is used to allow for await in the child component -->
         <Suspense>
           <UpdateContent :date="date"></UpdateContent>
@@ -20,14 +20,10 @@ import PageSectionTitle from '@/components/PageSectionTitle.vue'
 import UpdatesTable from '@/components/UpdatesTable.vue'
 import UpdateContent from '@/components/UpdateContent.vue'
 import { ref } from 'vue'
-import { useDisplay } from 'vuetify'
 import { useRoute } from 'vue-router'
 
 const route = useRoute()
 const { date } = route.params
-
-// mobile boolean for contitional rendering
-const { mdAndUp } = useDisplay()
 
 const title = ref('Updates')
 </script>