+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # ATLAS Website Changelog
 
+## [5.3.0]() (2026-08-31)
+
+- Improved website display on mobile
+- Updated contribution email to GitLab Service Desk address
+- Updated ATLAS data to [version 2026.08](https://github.com/mitre-atlas/atlas-data/blob/main/CHANGELOG.md#2026.08-2026-07-31)
+
 ## [5.2.0]() (2026-07-31)
 
 - Bundled ATLAS Navigator with website deployment